refactor(media): use kebab-case video viewer folders
