+++ frontend/src/features/video-viewer/components/sidebar/playlist-view/actions.js
