+++ frontend/src/features/video-viewer/components/sidebar/playlist-view/index.jsx
